Module Letsencrypt.Client

ACME Client.

This module provides client commands. Note: right now this module implements only the strict necessary in order to register an account, solve http-01 challenges provided by the CA, and fetch the certificate. This means that you will be able to maintain your server with this.

type t

type challenge =

| DNS

| HTTP

| ALPN

Scheduler monad

module type S = sig ... end

HTTP client interface

module type Client = sig ... end

module Solver (S : S) : sig ... end

module Make (S : S) (C : Client with type 'a t = 'a S.t) : sig ... end