Module 1-Http.Body

type t

The type of HTTP body.

val of_string : string -> t

of_string str makes a body from the given string str.

val to_string : t -> string Lwt.t

to_string body returns the full given body as a string.