Module 1-Http.Headers

type t

The type of HTTP headers.

val add : t -> string -> string -> t

add hdrs key value adds a key and a value to an existing hdrs headers.

val get : t -> string -> string option

get hdrs key retrieves a key from the given hdrs headers. If the header is one of the set of headers defined to have list values, then all of the values are concatenated into a single string separated by commas and returned. If it is a singleton header, then the first value is returned and no concatenation is performed.

val get_location : t -> Uri.t option

get_location hdrs is get hdrs "location".

val init_with : string -> string -> t

init_with key value constructs a fresh map of HTTP headers with a single key and value entry.

/

val to_string : t -> string