Module Client.Make

Parameters

module Http : HTTP_client.S

Signature

val initialise : ?ctx:Http.ctx -> endpoint:Uri.t -> ?email:string -> X509.Private_key.t -> (t, [> `Msg of string ]) Stdlib.result Lwt.t

initialise ~ctx ~endpoint ~email priv constructs a t by looking up the directory and account of priv at endpoint. If no account is registered yet, a new account is created with contact information of email. The terms of service are agreed on.

val sign_certificate : ?ctx:Http.ctx -> solver -> t -> (int -> unit Lwt.t) -> X509.Signing_request.t -> (X509.Certificate.t list, [> `Msg of string ]) Stdlib.result Lwt.t

sign_certificate ~ctx solver t sleep csr orders a certificate for the names in the signing request csr, and solves the requested challenges.